int/labels: inits to separate version string calculation for isolated testing
